+++ src/main/kotlin/io/github/woowacourse/archive/conversation/Conversation.kt
@@ -1,4 +1,4 @@
-package com.woowacourse.archive
+package io.github.woowacourse.archive.conversation
 
 import com.fasterxml.jackson.annotation.JsonBackReference
 import com.fasterxml.jackson.annotation.JsonManagedReference
