+++ src/main/kotlin/io/github/woowacourse/archive/conversation/ConversationController.kt
@@ -1,6 +1,6 @@
-package com.woowacourse.archive
+package io.github.woowacourse.archive.conversation
 
-import com.woowacourse.dto.ConversationDto
+import io.github.woowacourse.archive.dto.ConversationDto
 import org.springframework.http.ResponseEntity
 import org.springframework.web.bind.annotation.GetMapping
 import org.springframework.web.bind.annotation.PostMapping
