+++ src/main/kotlin/io/github/woowacourse/archive/conversation/ConversationRepository.kt
@@ -1,4 +1,4 @@
-package com.woowacourse.archive
+package io.github.woowacourse.archive.conversation
 
 import org.springframework.data.jpa.repository.JpaRepository
