+++ src/main/kotlin/io/github/woowacourse/archive/dto/ConversationDto.kt
@@ -1,7 +1,7 @@
-package com.woowacourse.dto
+package io.github.woowacourse.archive.dto
 
-import com.woowacourse.archive.Conversation
-import com.woowacourse.archive.Reply
+import io.github.woowacourse.archive.conversation.Conversation
+import io.github.woowacourse.archive.conversation.Reply
 import java.time.LocalDateTime
 
 data class ConversationDto(
