+++ src/main/kotlin/io/github/woowacourse/archive/slack/History.kt
@@ -1,4 +1,4 @@
-package com.woowacourse.slack
+package io.github.woowacourse.archive.slack
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties
 import com.fasterxml.jackson.annotation.JsonProperty
