+++ src/main/kotlin/io/github/woowacourse/archive/slack/Slack.kt
@@ -1,4 +1,4 @@
-package com.woowacourse.slack
+package io.github.woowacourse.archive.slack
 
 const val CONDITION_TRUE = "true"
