+++ src/test/kotlin/io/github/woowacourse/archive/conversation/ConversationServiceTest.kt
@@ -1,4 +1,4 @@
-package com.woowacourse.archive
+package io.github.woowacourse.archive.conversation
 
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
